requirements.txt
streamlit
supabase
python-dotenv
streamlit-geolocation
pillow